move files and update doc
