+++ .pre-commit-config.yaml
@@ -55,18 +55,18 @@ repos:
   rev: v2.2.5
   hooks:
   - id: codespell
-    args: [--toml=pyproject.toml]
+    args: [--toml, pyproject.toml]
     additional_dependencies: [tomli]
 
 - repo: https://github.com/psf/black
   rev: 23.7.0
   hooks:
   - id: black
-    args: [--config=pyproject.toml, --verbose]
+    args: [--config, pyproject.toml]
 
 - repo: https://github.com/pycqa/flake8
   rev: 6.0.0
   hooks:
   - id: flake8
-    args: [--verbose, --config=.flake8]
+    args: [--config, .flake8]
     additional_dependencies: [flake8-docstrings]
