+++ docs/source/usage_notes.rst
@@ -1,0 +1,11 @@
+.. _Usage :
+
+Usage Notes
+===========
+
+Command-Line Arguments
+----------------------
+.. argparse::
+   :prog: file_tree_check
+   :module: file_tree_check._parser
+   :func: pars_args
