+++ file_tree_check/_version.py
@@ -1,0 +1,6 @@
+# file generated by setuptools_scm
+# don't change, don't track in version control
+from __future__ import annotations
+
+__version__ = version = "0.1.dev91+g5838b20.d20230727"
+__version_tuple__ = version_tuple = (0, 1, "dev91", "g5838b20.d20230727")
