+++ file_tree_check/main.py
@@ -5,12 +5,13 @@ import logging
 import re
 from pathlib import Path
 
-from _parser import Parser
 from file_tree import FileTree
-from smartDirectoryPath import SmartDirectoryPath
-from smartFilePath import SmartFilePath
-from smartPath import SmartPath
-from statBuilder import StatBuilder
+
+from ._parser import Parser
+from .smartDirectoryPath import SmartDirectoryPath
+from .smartFilePath import SmartFilePath
+from .smartPath import SmartPath
+from .statBuilder import StatBuilder
 
 # Edit the following line to point to the config file location in your current installation:
 CONFIG_PATH = Path(__file__).parent / "config.ini"
